fix: change type of product's stock balance to decimal

+++ Choco.Backend.Api/Data/Models/Product.cs
@@ -6,7 +6,7 @@ public class Product : BaseModel
 {
     public required string Title { get; set; }
     public required ICollection<ProductTag> Tags { get; set; }
-    public required int StockBalance { get; set; }
+    public required decimal StockBalance { get; set; }
     public required bool IsBulk { get; set; }
     public required bool IsDeleted { get; set; }
     public ICollection<PriceHistory> Prices { get; set; }
